backend/requirements.txt
bentoml>=1.3,<1.4
setuptools<81
openai
langchain
langgraph
pydantic
python-dotenv
requests
